refactor: start splitting up cache.ts
